pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vedic-numerology-astrology"
version = "0.1.0"
description = "Computational Integration of Vedic Numerology and Sidereal Mechanics"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Norah Jones", email = "author@example.com"},
]
maintainers = [
    {name = "Norah Jones", email = "author@example.com"},
]
keywords = [
    "numerology",
    "astrology",
    "vedic",
    "sidereal",
    "swiss-ephemeris",
    "pyswisseph",
    "lahiri-ayanamsa",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Religion",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pytz>=2021.1",
    "PyYAML>=6.0",
    "suntime>=1.2.5",
]

[project.scripts]
vedic-numerology = "vedic_numerology.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
]
colab = [
    "plotly>=5.0.0",
    "ipywidgets>=7.6.0",
    "jupyter-widgets>=0.6.0",
]
ephemeris = [
    "pyswisseph>=2.08.00-1",
]

[project.urls]
Homepage = "https://github.com/astro-fusion/numerology-white-paper"
Documentation = "https://numerology-white-paper.readthedocs.io/"
Repository = "https://github.com/astro-fusion/numerology-white-paper"
"Bug Reports" = "https://github.com/astro-fusion/numerology-white-paper/issues"
"Source" = "https://github.com/astro-fusion/numerology-white-paper"

[tool.setuptools]
packages = ["vedic_numerology"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
vedic_numerology = ["config/*.yaml"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "swisseph.*",
    "flatlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=vedic_numerology --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "reference: marks tests that validate against reference cases",
]